Fix ULEB128 decoding
ULEB128 decoding for strings was broken the whole time and somehow I never noticed. I only noticed when /np threw errors and it seemed like the recipient was included in the message text, all in one string. The string even had 4 suspicious null bytes at the end (the sender ID lmfao). Turns out my noob to C# self decided to use BinaryReader Read() instead of ReadByte() and not think about why it returned an integer instead of a byte value